Add Support for Typelib Packaging (#33)
* Add support for installing girepository typelibs

This allows for apps that use other language bindings to be packaged
using linuxdeploy. For example, PyGObject uses the typelibs to
introspect bindings for Python.

* Add additional default pkg-config locations

Use default locations for Ubuntu in case libgtk-3-dev and librsvg2-dev are
not installed. A Python app wouldn't normally require these dependencies,
so use default values as a fallback during packaging for linuxdeploy.

* Fix failure if symbolic links already exist

* Fix gobject-introspection-1.0 library not found

The default typelib directory was incorrectly pointing to a location
for rpm based distros which was causing CI to fail with a library
not found error.

* Add additional dependencies to test typelibs

* Update dependencies for GTK4 to include typelibs